requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Werkzeug==3.0.1
scikit-learn==1.3.2
pandas==2.1.4
numpy==1.26.2
joblib==1.3.2
python-dotenv
requests
google-generativeai
pinecone
sentence-transformers
pymongo
gunicorn
groq
